Restructured the repository files
Repository files are now grouped in a repository directory in chirp.razor to reduce clutter in the file structure
