+++ src/Chirp.Razor/Repository/BaseRepository.cs
@@ -1,4 +1,4 @@
-namespace Chirp.Razor;
+namespace Chirp.Razor.Repository;
 
 public abstract class BaseRepository
 {
